Move TeamContributionCalendar into src
